Move Driver and CarGenerator to simulator package
