+++ src/main/scala/pl/edu/pw/elka/Main.java
@@ -3,10 +3,10 @@ package pl.edu.pw.elka;
 import java.util.Random;
 import java.util.Timer;
 import pl.edu.pw.elka.database.Database;
-import pl.edu.pw.elka.database.Driver;
+import pl.edu.pw.elka.simulator.Driver;
 import pl.edu.pw.elka.drawer.TerrainDrawer;
 import pl.edu.pw.elka.enums.Roads;
-import pl.edu.pw.elka.generator.CarGenerator;
+import pl.edu.pw.elka.simulator.CarGenerator;
 import pl.edu.pw.elka.knowledgeDatabase.Junction;
 import pl.edu.pw.elka.terrain.Terrain;
 import pl.edu.pw.elka.terrain.XJunction;
